refactor: clean up helper methods
- remove the unlinkFile method
- remove all awaits when calling guild config read/write methods
- fix a bug (introduced by the previous commit) where an error would occur when using /random-channel draw without being in a voice channel

+++ events/messageCreate.js
@@ -19,31 +19,32 @@ module.exports = {
       }
 
       // if message author (as a guild member) has the slowmode role
-      if (message?.member.roles.cache.get(guildConfig.slowmodeRoleId)) {
-        // if channel is affected by slowmode role
-        if (guildConfig.slowmodeChannels.includes(message.channelId) && message.channel.type === 0) {
-          if (!latestMessages.has(message.channelId)) {
-            latestMessages.set(message.channelId, new Collection());
-          }
-          if (!latestMessages.get(message.channelId).has(message.author.id)) {
-            latestMessages.get(message.channelId).set(message.author.id, message.createdTimestamp);
-          } else if (message.createdTimestamp - latestMessages.get(message.channelId).get(message.author.id) < guildConfig.slowmodeDelay * 1000) {
-            const t1 = Date.now();
-            await message.delete();
-            const t2 = Date.now();
-            logAction({ name: `${this.name} event handling - applying slowmode`, guild: message.guild, message: message });
+      if (
+        message?.member.roles.cache.get(guildConfig.slowmodeRoleId)
+        && guildConfig.slowmodeChannels.includes(message.channelId)
+        && message.channel.type === 0
+      ) {
+        if (!latestMessages.has(message.channelId)) {
+          latestMessages.set(message.channelId, new Collection());
+        }
+        if (!latestMessages.get(message.channelId).has(message.author.id)) {
+          latestMessages.get(message.channelId).set(message.author.id, message.createdTimestamp);
+        } else if (message.createdTimestamp - latestMessages.get(message.channelId).get(message.author.id) < guildConfig.slowmodeDelay * 1000) {
+          const t1 = Date.now();
+          await message.delete();
+          const t2 = Date.now();
+          logAction({ name: `${this.name} event handling - applying slowmode`, guild: message.guild, message: message });
 
-            if (t2 - t1 > 1000 && !guildConfig.missingPermissions.includes("ModerateMembers")) {
-              // if the delete method takes too much time, it means that the user is spamming too quickly and Roberto reaches their API rate limit
-              // in that case, timeout the user for 2x the slowmode delay
-              // (of course, an admin can still revoke the timeout before it expires)
-              const timeoutMessage = `${guildConfig.slowmodeDelay * 4} seconds timeout for spamming`;
-              message.member.timeout(guildConfig.slowmodeDelay * 4000, timeoutMessage);
-              logAction({ name: `${this.name} event handling - applying a timeout`, guild: message.guild, member: message.member, reason: timeoutMessage });
-            }
-          } else {
-            latestMessages.get(message.channelId).set(message.author.id, message.createdTimestamp);
+          if (t2 - t1 > 1000 && !guildConfig.missingPermissions.includes("ModerateMembers")) {
+            // if the delete method takes too much time, it means that the user is spamming too quickly and Roberto reaches their API rate limit
+            // in that case, timeout the user for 2x the slowmode delay
+            // (of course, an admin can still revoke the timeout before it expires)
+            const timeoutMessage = `${guildConfig.slowmodeDelay * 4} seconds timeout for spamming`;
+            message.member.timeout(guildConfig.slowmodeDelay * 4000, timeoutMessage);
+            logAction({ name: `${this.name} event handling - applying a timeout`, guild: message.guild, member: message.member, reason: timeoutMessage });
           }
+        } else {
+          latestMessages.get(message.channelId).set(message.author.id, message.createdTimestamp);
         }
       }
     } catch (err) {
